+++ R/set_blanket.R
@@ -12,11 +12,15 @@
 #'
 #' @param theme A ggplot2 theme (e.g. [theme_lighter()] or [theme_darker()]).
 #' @param ... Provided to require argument naming, support trailing commas etc.
-#' @param colour A default hex code for the colour of most geoms. Defaults to blue.
-#' @param fill A default hex code for the fill of most geoms. Inherits from colour.
+#' @param col A default hex code for the colour and fill of most geoms. Defaults to "#357BA2FF" (i.e. `blue`).
+#' @param colour A default hex code for the colour of most geoms. Defaults to col.
+#' @param fill A default hex code for the fill of most geoms. Defaults to col.
 #' @param size A default size for the point geom. Defaults to 1.5. The pointrange size defaults to dividing by 6 (i.e. 0.25).
 #' @param linewidth A default linewidth for most geoms. Defaults to 0.66.
-#' @param bordertype A default linetype for geoms with unnecessary border lines. Defaults to 0.
+#' @param linetype A default linetype for geoms not specified by linetype_border, linetype_box. or linetype_sf. Defaults to 1.
+#' @param linetype_sf A default linetype for sf geoms. Defaults to 0.
+#' @param linetype_box A default linetype for boxplot and crossbar geoms. Defaults to 1.
+#' @param linetype_border A default linetype for polygon-is geoms (other than boxplot, crossbar or sf). Defaults to 0.
 #' @param col_palette_discrete For a discrete scale, a character vector of hex codes.
 #' @param col_palette_continuous For a continuous scale, a character vector of hex codes.
 #' @param col_palette_o For an ordinal scale, a `scales::pal_*()` function.
@@ -58,11 +62,16 @@
 set_blanket <- function(
     theme = theme_lighter(),
     ...,
-    colour = "#357BA2FF",
-    fill = colour,
+    col = "#357BA2FF",
+    colour = col,
+    fill = col,
     linewidth = 0.66,
-    bordertype = 0,
     size = 1.5,
+    linetype = 1,
+    linetype_border = 0,
+    linetype_box = 1,
+    linetype_sf = 1,
+
     col_palette_discrete = jumble,
     col_palette_continuous = viridisLite::mako(n = 9, direction = -1),
 
@@ -86,10 +95,18 @@ set_blanket <- function(
     panel_grid_transparent = panel_grid_transparent
   )
 
-  weave_geom_colour_fill(colour = colour, fill = fill)
-  weave_geom_linewidth(linewidth = linewidth)
-  weave_geom_bordertype(bordertype = bordertype)
+  weave_geom_col(col = col, colour = colour, fill = fill)
+
   weave_geom_size(size = size)
+
+  weave_geom_linewidth(linewidth = linewidth)
+
+  weave_geom_linetype(
+    linetype = linetype,
+    linetype_border = linetype_border,
+    linetype_box = linetype_box,
+    linetype_sf = linetype_sf
+  )
 
   weave_geom_text()
   weave_geom_label()