+++ tests/testthat/test-article-go-further.R
@@ -323,3 +323,5 @@ test_that(test_name, {
 
   vdiffr::expect_doppelganger(test_name, p)
 })
+
+set_blanket()