+++ tests/testthat/test-datetime.R
@@ -58,3 +58,5 @@ test_that(test_name, {
 
   vdiffr::expect_doppelganger(test_name, p)
 })
+
+set_blanket()